// src/app/api/model/accountRead.ts
/**
 * FinanzApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountTypeEnum } from './accountTypeEnum';


/**
 * Representación pública de una cuenta
 */
export interface AccountRead { 
    id: string;
    group_id: string;
    name: string;
    type: AccountTypeEnum;
    opening_balance: number;
    balance: number;
    currency: string;
    color: string | null;
    icon: string | null;
    is_active: boolean;
    created_by: string | null;
    updated_by: string | null;
    created_at: string;
    updated_at: string;
}
export namespace AccountRead {
}