// src/app/api/model/createTransactionRequest.ts
/**
 * FinanzApp API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface CreateTransactionRequest {
  type: TransactionTypeEnum;
  amount: number;
  category_id?: string | null;
  to_account_id?: string | null;
  date: string;
  notes?: string | null;
}
export namespace CreateTransactionRequest {}